wxGUI/gmodeler: refactor - create model_convert.py
move model -> script converters there
